sunset_predictor/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
requests>=2.31.0
astral>=3.2
tqdm>=4.65.0
tensorboard>=2.13.0

